+++ RELEASENOTES.md
@@ -4,3 +4,6 @@
 * Added workaround fix when using camera zoom setting with value other than 50 in MSFS general options will cause Pop Out Panel Manager pop out to fail. This is an existing MSFS bug where saving and loading of custom camera view is currently broken for zoom level other than 50.
 
 * Fixed issue where full screen mode for pop out panel does not work on certain aircraft configuration.
+
+* Updated logic to verify custom camera view has been loaded before starting pop out process.
+